pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "intersectionqa"
version = "0.1.0"
description = "Code-only CAD spatial reasoning benchmark tooling."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
intersectionqa-generate = "scripts.generate_dataset:main"
intersectionqa-validate = "scripts.validate_dataset:main"
intersectionqa-aabb-baseline = "scripts.evaluate_baseline:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
